consertando erros do exercicio
