--- para_o_lar/src/controllers/musicController.js
@@ -1,1 +1,0 @@
-
